Initialization code
Currently, there is code that handles:
  - Creation and configuration of the tun device file descriptor.
  - Binding to a UDP Socket

NOTE: Right now, we are binding on a UDP socket. This will have to
be changed to a TCP socket.

There are also stubs for tx_loop and rx_loop, which are the entry
points for two threads, one for sending and the other for receiving.